+++ src/main/app/Resources/modules/content/form/components/data.jsx
@@ -1,47 +1,17 @@
-import React, {createElement} from 'react'
+import React from 'react'
 import {PropTypes as T} from 'prop-types'
-import classes from 'classnames'
 import get from 'lodash/get'
 import isEmpty from 'lodash/isEmpty'
 
 import {trans} from '#/main/app/intl/translation'
-import {toKey} from '#/main/core/scaffolding/text'
 import {Button} from '#/main/app/action/components/button'
-import {Toolbar} from '#/main/app/action/components/toolbar'
 import {MENU_BUTTON, CALLBACK_BUTTON} from '#/main/app/buttons'
-import {ContentTitle} from '#/main/app/content/components/title'
 import {ContentMeta} from '#/main/app/content/components/meta'
-import {Form} from '#/main/app/content/form/components/form' // TODO : use container instead
-import {FormFieldset} from '#/main/app/content/form/components/fieldset'
-import {FormSections, FormSection} from '#/main/app/content/form/components/sections'
+import {Form} from '#/main/app/content/form/containers/form'
 
 import {constants} from '#/main/app/content/form/constants'
-import {createFormDefinition} from '#/main/app/content/form/utils'
 import {DataFormSection as DataFormSectionTypes} from '#/main/app/content/form/prop-types'
-
-function getSectionId(section, formId = null) {
-  let id = formId ? `${formId}-` : ''
-
-  id += section.id || toKey(section.title)
-
-  return id
-}
-
-function getSectionErrors(sectionFields = [], errors = {}) {
-  let sectionErrors = []
-
-  sectionFields.map(field => {
-    if (get(errors, field.name)) {
-      sectionErrors.push(get(errors, field.name))
-    }
-
-    if (field.linked) {
-      sectionErrors = sectionErrors.concat(getSectionErrors(field.linked, errors))
-    }
-  })
-
-  return sectionErrors
-}
+import {FormContent} from '#/main/app/content/form/containers/content'
 
 const FormModes = props =>
   <div className="form-mode">
@@ -72,27 +42,13 @@ FormModes.propTypes = {
 }
 
 const FormData = (props) => {
-  const hLevel = props.level + (props.title ? 1 : 0)
-  let hDisplay
-  if (props.displayLevel) {
-    hDisplay = props.displayLevel + (props.title ? 1 : 0)
-  }
-
-  const sections = createFormDefinition(props.mode, props.definition || props.sections, props.locked, props.data)
-
-  const primarySections = 1 === sections.length ? [sections[0]] : sections.filter(section => section.primary)
-  const otherSections = 1 !== sections.length ? sections.filter(section => !section.primary) : []
-  let openedSection = otherSections.find(section => section.defaultOpened)
-
   const disabled = typeof props.disabled === 'function' ? props.disabled(props.data) : props.disabled
-
-  if (props.autoFocus && !isEmpty(primarySections) && !isEmpty(primarySections[0].fields)) {
-    primarySections[0].fields[0].autoFocus = true
-  }
 
   return (
     <Form
       id={props.id}
+      name={props.name}
+      dataPart={props.dataPart}
       className={props.className}
       embedded={props.embedded}
       flush={props.flush}
@@ -100,15 +56,12 @@ const FormData = (props) => {
       level={props.level}
       displayLevel={props.displayLevel}
       title={props.title}
-      errors={!isEmpty(props.errors)}
-      validating={props.validating}
-      pendingChanges={props.pendingChanges}
       alertExit={props.alertExit}
+      lock={props.lock}
       save={props.save}
       cancel={props.cancel}
-      lock={props.lock}
-      getLock={props.getLock}
-      unlock={props.unlock}
+      buttons={props.buttons}
+      target={props.target}
     >
       {props.meta &&
         <ContentMeta
@@ -125,87 +78,18 @@ const FormData = (props) => {
         />
       }
 
-      {primarySections.map(primarySection =>
-        <section
-          id={`${getSectionId(primarySection, props.id)}-section`}
-          key={primarySection.id || toKey(primarySection.title)}
-          className={classes('form-primary-section', primarySection.className, !props.flush && 'mb-3')}
-        >
-          <ContentTitle
-            level={hLevel}
-            displayed={false}
-            title={primarySection.title}
-            subtitle={primarySection.subtitle}
-          />
-
-          {!isEmpty(primarySection.actions) &&
-            <Toolbar
-              id={`${getSectionId(primarySection, props.id)}-actions`}
-              buttonName="btn"
-              className="text-right form-group"
-              size="sm"
-              actions={primarySection.actions}
-            />
-          }
-
-          <FormFieldset
-            id={getSectionId(primarySection, props.id)}
-            mode={props.mode}
-            disabled={props.disabled || primarySection.disabled}
-            fields={primarySection.fields}
-            data={props.data}
-            errors={props.errors}
-            help={primarySection.help}
-            validating={props.validating}
-            updateProp={props.updateProp}
-            setErrors={props.setErrors}
-          >
-            {primarySection.component && createElement(primarySection.component)}
-            {!primarySection.component && primarySection.render && primarySection.render(props.data, props.errors)}
-          </FormFieldset>
-        </section>
-      )}
-
-      {0 !== otherSections.length &&
-        <FormSections
-          level={hLevel}
-          displayLevel={hDisplay}
-          defaultOpened={openedSection ? getSectionId(openedSection, props.id) : undefined}
-          flush={props.flush}
-          className={classes(!props.flush && 'mb-3')}
-        >
-          {otherSections.map(section => (
-            <FormSection
-              id={getSectionId(section, props.id)}
-              className={section.className}
-              key={getSectionId(section, props.id)}
-              icon={section.icon}
-              title={section.title}
-              subtitle={section.subtitle}
-              errors={getSectionErrors(section.fields, props.errors)}
-              validating={props.validating}
-              actions={section.actions}
-              fill={section.fill}
-            >
-              <FormFieldset
-                id={`${getSectionId(section, props.id)}-fieldset`}
-                mode={props.mode}
-                disabled={props.disabled || (typeof section.disabled === 'function' ? section.disabled(props.data) : section.disabled)}
-                fields={section.fields}
-                data={props.data}
-                errors={props.errors}
-                help={section.help}
-                validating={props.validating}
-                updateProp={props.updateProp}
-                setErrors={props.setErrors}
-              >
-                {section.component && createElement(section.component)}
-                {!section.component && section.render && section.render(props.data, props.errors)}
-              </FormFieldset>
-            </FormSection>
-          ))}
-        </FormSections>
-      }
+      <FormContent
+        id={props.id}
+        name={props.name}
+        dataPart={props.dataPart}
+        level={props.level}
+        displayLevel={props.displayLevel}
+        flush={props.flush}
+        autoFocus={props.autoFocus}
+        disabled={disabled}
+        definition={props.definition || props.sections}
+        locked={props.locked}
+      />
 
       {props.children}
     </Form>
@@ -213,7 +97,7 @@ const FormData = (props) => {
 }
 
 FormData.propTypes = {
-  id: T.string.isRequired,
+  id: T.string,
   /**
    * Is the form embed into another ?
    *
@@ -228,9 +112,6 @@ FormData.propTypes = {
   className: T.string,
   mode: T.string.isRequired,
   disabled: T.oneOfType([T.bool, T.func]),
-  errors: T.object,
-  validating: T.bool,
-  pendingChanges: T.bool,
   /**
    * Alerts the user when leaving the form with unsaved changes
    */
@@ -254,9 +135,6 @@ FormData.propTypes = {
     className: T.string.isRequired
   }),
 
-  getLock: T.func,
-  unlock: T.func,
-
   /**
    * The save action of the form.
    */
@@ -275,8 +153,6 @@ FormData.propTypes = {
     // todo find a way to document custom action type props
   }),
   setMode: T.func.isRequired,
-  setErrors: T.func.isRequired,
-  updateProp: T.func.isRequired,
   children: T.node
 }
